Change Four: Reorder Wishlist Items
This change builds on the design decisions made in Change Three to put
Items as items in a Map within the Wishlist, and to split out a list
of ids to track the ordering of the Items. Now we have the power to
reorder the items in the Wishlist.

From the perspective of the client of this service, reordering an Item
seemed like a distinct business rule, which is why the service has an
explicit "reorderWishlistItem" function.

However, from the perspective of the Repository, it seemed like
updating the Item's position was just one possible update we could make
on the persisted Item state. That is why the Repository supports a more
generic "updateItem" function. It only supports updating the index, but
it is open to extension to updating other attributes as well. Maybe a
violation of YAGNI -- perhaps in non-demo code, it would have been
better to write a "reorderItem" function, and defer a generic "update"
function later (and refactor the reorderItem function to bundle into
the update function). For the demo, I wanted to demonstrate that the
Repository doesn't always have to have 1:1 functionality with the biz
logic defined in the domain use case.